notification-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.1.1
pymongo==4.3.3
pydantic==2.4.2
pika==1.3.2
python-dotenv==1.0.1
python-consul==1.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
aiosmtplib==2.0.2 